Fix potential integer overflow (#182)
When injecting a DLL into a process, it is possible that the process memory has been corrupted.  The values in the import table for the process could be incorrect, which could cause an integer overflow when calculating the size of the new import table.  Add code to protect against this to UPDATE_IMPORTS_XX.