CI: Disable ARM32 in the CI builds. (#342)
The 32-bit ARM builds no longer succeed with the latest platform SDK because 32-bit ARM support has been deprecated and there are no longer libs for things like ws_32. This is why the windows-2022 CI builds are failing for the two 32-bit ARM jobs. 32-bit ARM isn't much of a concern for anybody, I'm just going ahead and disabling it for the CI runs for now.

Note that this is not removing 32-bit ARM support from Detours, just (temporarily) from the CI builds.